backend/requirements.txt
scikit-build
fastapi
uvicorn
langchain
langchain_experimental
langchain_openai
langchain-community
huggingface_hub[cli,torch,tensorflow]
transformers
torch
tokenizers
sentence_transformers
pydantic-settings
tiktoken
datasets
ollama
gguf
gptcache
jinja2
pymongo
starlette_context
sse-starlette
pyjwt
cryptography
python-multipart
